Cache for pictures
Recent Improvements:
   . Zip files contain folder structure for Data folder. without that, no posters... However, even if you re-create the category and re-scan, no need to download posters again, way faster
   . Just work with no fuss on folders names in Kanjis [Sparanoid thanks for reporting, it never worked before but that made me want to do it]
   . Use the syntax "aid:2656:アフロサムライ" or "aid:2656" and selecting "ja" instead of "main" as first language in the agent settings, will make the title stays in Kanji. Beware of caching however, as changing the language can retain the previous title.... you need to delete all, empty trash, clean bundles, restart scanning from scratch
   . Local themes and poster cache, once downloaded, never gets downloaded again
   . Allow for replacement of mp3 themes
   . Enables all languages and synonym lookup, and strangely seem faster since it recognize way more series in seconds
   . AniDB posters can be enabled or disabled from the agent settings window, just like the TVDB posters, banners, and the language selection for the title and episodes (ja in in Kanji, x-jat is in Romaji, main is the main title in anidb)
   . 2 week cache for XML files. can use aid:xxxxx to force anime if not detected. Do we need to restart the agent afterwards to cache again? i don't know yet...
   . average length working back.
   
A lot changed in a week, thanks to AtomicStrawberry giving me back a bit of my apetite for programming and by few peoples that just reported issues, like the Kanji not working... Not aware of other agents with that working :D

I don't recommend downloading the AniDB cover since it's rarely there and not normalised
Please let me know if you like this one, and if it is faster that before and any other agent. the recognition rate should have jumped significantly since it checks synonyms and all languages unlike before. Please feedback, as it is only available here, also improvement requests are welcome. The code is on github