EndOfLine369:tvdbOverview used if no AniDB summary
Noticed this on new series and some AniDB neglected series. So why not use the TVDB summary if successfully pulled earlier if the AniDB is blank so at least data is there. And when AniDB gets updated, it will replace it with the AniDB summary on the next metadata update.
https://github.com/ZeroQI/Hama.bundle/issues/33 fix by EndOfLine369