Reduce incorrect 's0' mappingLlist['TVDB'] entries
Some tvdb series contain multiple AniDB series in the TVDB season 0 as specials. If the AniDB entry is 'defaulttvdbseason' == 0, even if there is specific mapping, it adds all series as 's0' in the TVDB dict. This plays bad when trying to find the anidb id from the tvdb number. So only let mapping that has no special 'season 1' entries be entered under the generic 's0' key.