Fix AniDB_array generation for absoute numbering
If tvdb3 was being used, the entry would not be seen in the 'anidb_ep' call as was an absolute number. Convert back to normap episode number from 'absolute_map' data then call 'anidb_ep'.

If tvdb4 was being used, the AniDB id cannot be found as it is not using TVDB mapping. So all AniDBs will need to be pulled.