Remove AniDB module logic from 'LoadFile()'
The only true AniDB logic left is the sleep pause. Added that in as a function argument 'sleep' with default of '0'. So AniDB will just be overriding that to '6' ag invocation.

The 'netLocked' logic was not really AniDB specic to has been made generic and applied to the function in general just as the 'Thread.Lock()' is.