Reduce AniDB xml pulls via mappingList['possible_anidb3']
Reduces the AniDB xml pulls for anidb3 setups. Since it is translated and added in as 'tvdb', it could also possibly be a plain 'tvdb' entry when seen by HAMA. So we only want to pull all AniDB xmls on a possible 'anidb3' entry. So 'possible_anidb3' is defined if an episode number of >100 is found which is what ASS will add it in as.