Fix 'is_primary_series' if there is direct s1e1 mapping
#392
----------
In addition to https://github.com/ZeroQI/Hama.bundle/commit/8581975837a48f9ce3e045b7ce0d568bfdc50037
----------
If there is any s1->s0/s1 mapping seen, it is considered to not be the primary AniDB series entry. But in the case of 'anidb-10538', it has mapping that is actually remapping directly to s1e1. So an additional check needs to be put in to see if this direct mapping exists and put into the test for 'is_primary_series'.
----------
<mapping anidbseason="1" tvdbseason="1">;2-1;3-2;4-3;5-4;6-5;7-6;8-7;9-8;10-9;11-10;12-11;13-12;</mapping>
----------
[+] AniDBid: 10538, TVDBid: 281275, defaulttvdbseason:   1, offset:   0, name: Saenai Heroine no Sodatekata
    - season: [ 1],           [ 0], range:       [000-000], offset:   0, text: 1-1
    - season: [ 1],           [ 1], range:       [000-000], offset:   0, text: 2-1;3-2;4-3;5-4;6-5;7-6;8-7;9-8;10-9;11-10;12-11;13-12
             -----          ------
             10538          281275