Fix the setting of TMDB/IMDB entries
Issue: If a TVDB source id, it would complete the loop of all entries. This would then cause it to just take the TMDB/IMDB values from the last entry in the list. Also, if IMDB source id was used, it was not being looked for.

Fix:
Add the 'IMDB_id' to the forcedID list to match.
Set the values to be set for the single entry found if either  AniDB/TMDB/IMDB is matched as there will only be one and break the loop.
Set the TMDB/IMDB values when the primary AniDB entry is found for TheTVDB/multi-season numbered series.
Since the TMDB/IMDB are now being set in the loop, only the check for if it was found needs to be in the 'else' post loop section.