Adjust image priority per rating and main image
This was accidently lost when converted to "common.poster_rank()". As images are sorted by rating average, just add the image count to the rank using "poster_rank()" already included "rank_adjustment" parameter. And if the image is the primary one assigned to the main entry, set it to the highest priority (0/no adjustment).