Update AniDB.py
Some of the entries in my library had urls in it that escaped sanitization. After a bit of digging, it seems that the issue stems from those entries using the 'full' link in the summary, rather than the short version (e.g. "[https://anidb.net/character/270](https://anidb.net/character/270)" rather than "[https://anidb.net/ch270](https://anidb.net/ch270)".

I did a few tests and this seems to work for fixing the issue.


A more aggressive solution would be to simply match until the first whitespace character. Replacing both link-cleaning lines with:
```python
  summary = re.sub(r'https?://anidb\.net/[\S]+ \[(?P<text>.+?)\]', r'\g<text>', summary)             # Replace all links
```

This seems to work fine for the titles I've tested it with, but I wasn't sure if there was a reason for the precision approach, so I opted for a similarly precise approach in the commit.